requirements.txt
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
coloredlogs==14.0
cycler==0.10.0
freetype-py==2.1.0.post1
humanfriendly==8.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
Jinja2==2.11.2
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mkl-fft==1.1.0
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
Pillow==7.1.2
pyaml==20.4.0
pycairo==1.19.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
scipy==1.4.1
svgpathtools==1.3.3
svgwrite==1.4
tensorboard==2.2.2
torch==1.3.1
torch-tools==0.1.0
torchfile==0.1.0
torchvision==0.4.2
tornado==6.0.4
tqdm==4.46.1
urllib3==1.25.9
visdom==0.1.8.9
websocket-client==0.57.0
